Adicionando exercicios de python

+++ _aulas/aula007c.py
@@ -1,4 +1,4 @@
-n1 = int(input('Digite uma nota: '))
-n2 = int(input('Digite outra nota: '))
+n1 = float(input('Digite uma nota: '))
+n2 = float(input('Digite outra nota: '))
 med = (n1 + n2) / 2
 print('A média das notas é: {}'.format(med))
